deployment/requirements.txt
streamlit
pandas
pillow
matplotlib
numpy
tensorflow == 2.15.1
